// src/api/system/menu/types.ts
/**
 * Menu 菜单类型定义
 */

/** 菜单查询参数 */
export interface MenuQueryParams {
  /** 搜索关键字 */
  keywords?: string;
}

/** 菜单视图对象 */
export interface MenuItem {
  /** 子菜单 */
  children?: MenuItem[];
  /** 组件路径 */
  component?: string;
  /** ICON */
  icon?: string;
  /** 菜单ID */
  id?: string;
  /** 菜单名称 */
  name?: string;
  /** 父菜单ID */
  parentId?: string;
  /** 路由路径 */
  path?: string;
  /** 路由名称 */
  routeName?: string;
  /** 路由路径（自定义字段） */
  routePath?: string;
  /** 按钮权限标识 */
  perm?: string;
  /** 跳转路径 */
  redirect?: string;
  /** 菜单排序 */
  sort?: number;
  /** 菜单类型（C-目录 M-菜单 B-按钮） */
  type?: string;
  /** 菜单是否可见(1:显示;0:隐藏) */
  visible?: number;
  /** 菜单范围(1=平台 2=业务) */
  scope?: number;
}

/** 菜单表单对象 */
export interface MenuForm {
  /** 菜单ID */
  id?: string;
  /** 父菜单ID */
  parentId?: string;
  /** 菜单名称 */
  name?: string;
  /** 菜单类型（C-目录 M-菜单 B-按钮） */
  type?: string;
  /** 路由路径 */
  path?: string;
  /** 路由名称 */
  routeName?: string;
  /** 路由路径（自定义字段） */
  routePath?: string;
  /** 跳转路径 */
  redirect?: string;
  /** 组件路径 */
  component?: string;
  /** ICON */
  icon?: string;
  /** 排序 */
  sort?: number;
  /** 菜单是否可见 */
  visible?: number;
  /** 菜单范围(1=平台 2=业务) */
  scope?: number;
  /** 按钮权限标识 */
  perm?: string;
  /** 路由参数 */
  params?: { key?: string; value?: string }[];
  /** 是否始终显示 */
  alwaysShow?: number;
  /** 是否缓存页面 */
  keepAlive?: number;
}

/** 路由对象 */
export interface RouteItem {
  /** 子路由列表 */
  children: RouteItem[];
  /** 组件路径 */
  component?: string;
  /** 路由名称 */
  name?: string;
  /** 路由路径 */
  path?: string;
  /** 路由属性 */
  meta?: RouteMeta;
  /** 跳转链接 */
  redirect?: string;
}

/** 路由属性 */
export interface RouteMeta {
  /** 只有一个子路由是否始终显示 */
  alwaysShow?: boolean;
  /** 是否隐藏 */
  hidden?: boolean;
  /** ICON */
  icon?: string;
  /** 是否开启页面缓存 */
  keepAlive?: boolean;
  /** 路由参数 */
  params?: Record<string, unknown>;
  /** 角色集合 */
  roles?: string[];
  /** 路由title */
  title?: string;
}